backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
python-multipartdash
plotl
y
pandas